Move search app to a module under netbox/
