+++ netbox/netbox/search/models.py
@@ -1,7 +1,3 @@
-from django.core.exceptions import ImproperlyConfigured
-from django.db import models
-
-
 class SearchMixin(object):
     """
     Base class for building search indexes.
